Everything seems to be merged and working
