+++ gemdaq-testing/gemhardware/include/gem/hw/GEMHwDevice.h
@@ -8,6 +8,12 @@
 
 #include "uhal/uhal.hpp"
 
+#define DEBUG(MSG) LOG4CPLUS_DEBUG(logGEMHw_ , MSG)
+#define INFO(MSG)  LOG4CPLUS_INFO(logGEMHw_  , MSG)
+#define WARN(MSG)  LOG4CPLUS_WARN(logGEMHw_  , MSG)
+#define ERROR(MSG) LOG4CPLUS_ERROR(logGEMHw_ , MSG)
+#define FATAL(MSG) LOG4CPLUS_FATAL(logGEMHw_ , MSG)
+
 #define MAX_VFAT_RETRIES 25
 
 typedef uhal::exception::exception uhalException;
@@ -62,12 +68,6 @@ namespace gem {
 	
 	bool isGEMHwDeviceConnected() { return gemHWP_ != 0; };
 	
-	virtual std::string getBoardID()   ;
-	virtual std::string getSystemID()  ;
-	//virtual std::string getIPAddress() ; //need to implement!!!
-	virtual std::string getFirmwareVer( std::string const& fwRegPrefix="glib_regs.sysregs") ;
-	virtual std::string getFirmwareDate(std::string const& fwRegPrefix="glib_regs.sysregs") ;
-
 	/**
 	 *Generic read/write functions or IPBus devices
 	 * operation will be the same for the GLIB, MP7, VFAT2/3, 
@@ -95,11 +95,15 @@ namespace gem {
 
 	//perform a single write transaction
 	virtual void     writeReg( std::string const& regName, uint32_t const);
+	virtual void     writeReg( std::string const& regPrefix,
+				   std::string const& regName, uint32_t const val) {
+	  std::string name = regPrefix+"."+regName;
+	  return writeReg(name, val); };
 	//write list of registers in a single transaction (one dispatch call) using the supplied vector regList
 	virtual void     writeRegs(std::vector<std::pair<std::string, uint32_t> > const& regList);
 	//write single value to a list of registers in a single transaction (one dispatch call) using the supplied vector regList
 	virtual void     writeValueToRegs(std::vector<std::string> const& regList, uint32_t const& regValue);
-
+	
 	//write zero to a single register
 	virtual void     zeroReg(  std::string const& regName) { writeReg(regName,0); };
 	//write zero to a list of registers in a single transaction (one dispatch call) using the supplied vector regNames
@@ -107,11 +111,10 @@ namespace gem {
 
 	virtual std::vector<uint32_t> readBlock( std::string const& regName);
 	virtual std::vector<uint32_t> readBlock( std::string const& regName,
-					 size_t const nWords);
+						 size_t const nWords);
 	virtual void writeBlock(std::string const& regName,
 				std::vector<uint32_t> const values);
-	virtual void zeroBlock( std::string const& regName) {
-	  writeReg(regName, 0); }
+	virtual void zeroBlock( std::string const& regName) { writeReg(regName, 0); }
 
 
 	// These methods provide access to the member variables
@@ -134,13 +137,13 @@ namespace gem {
 	void setDeviceID(std::string const& deviceID) {
 	  deviceID_ = deviceID; };
 	
-	uhal::HwInterface& getGEMHwInterface();// ;
+	uhal::HwInterface& getGEMHwInterface() const;
 	
 	void updateErrorCounters(std::string const& errCode);
 	
 	DeviceErrors ipBusErrs;
 	
-	std::string printErrorCounts();
+	std::string printErrorCounts() const;
 	
       protected:
 	uhal::ConnectionManager *gemConnectionManager;
@@ -154,9 +157,9 @@ namespace gem {
 	std::string deviceIPAddr_;
 	std::string deviceID_;
 		
-	//std::string registerToChar(uint32_t value);
+	//std::string registerToChar(uint32_t value) const;
 	
-	std::string uint32ToString(uint32_t const val) {
+	std::string uint32ToString(uint32_t const val) const{
 	  std::stringstream res;
 	  res << char((val & uint32_t(0xff000000)) / 16777216);
 	  res << char((val & uint32_t(0x00ff0000)) / 65536);
