+++ gemdaq-testing/gemhardware/src/common/HwVFAT2.cc
@@ -2,11 +2,11 @@
 
 #include "gem/hw/vfat/HwVFAT2.h"
 
-#define DEBUG(MSG) LOG4CPLUS_DEBUG(logGEMHw_ , MSG)
-#define INFO( MSG) LOG4CPLUS_INFO( logGEMHw_ , MSG)
-#define WARN( MSG) LOG4CPLUS_WARN( logGEMHw_ , MSG)
-#define ERROR(MSG) LOG4CPLUS_ERROR(logGEMHw_ , MSG)
-#define FATAL(MSG) LOG4CPLUS_FATAL(logGEMHw_ , MSG)
+//#define DEBUG(MSG) LOG4CPLUS_DEBUG(logGEMHw_ , MSG)
+//#define INFO( MSG) LOG4CPLUS_INFO( logGEMHw_ , MSG)
+//#define WARN( MSG) LOG4CPLUS_WARN( logGEMHw_ , MSG)
+//#define ERROR(MSG) LOG4CPLUS_ERROR(logGEMHw_ , MSG)
+//#define FATAL(MSG) LOG4CPLUS_FATAL(logGEMHw_ , MSG)
 
 gem::hw::vfat::HwVFAT2::HwVFAT2(xdaq::Application* vfatApp,
 				std::string const& vfatDevice):
