+++ gemdaq-testing/gemhardware/xml/GEMController.xml
@@ -7,9 +7,9 @@
     <properties xmlns="urn:xdaq-application:gem::hw::vfat::GEMController"
 		  xsi:type="soapenc:Struct">
 	<device xsi:type="xsd:string">VFAT9</device>
-	<settingsFile xsi:type="xsd:string">/home/mdalchen/test/gem-daq-code/gemdaq-testing/gemhardware/xml/gem_conf.xml</settingsFile>
+	<settingsFile xsi:type="xsd:string">${BUILD_HOME}/gemdaq-testing/gemhardware/xml/gem_conf.xml</settingsFile>
       </properties>
     </xc:Application>
-    <xc:Module>/home/mdalchen/test/gem-daq-code/gemdaq-testing/gemhardware/lib/linux/x86_64_slc6/libgem_hw.so</xc:Module>
+    <xc:Module>${BUILD_HOME}/gemdaq-testing/gemhardware/lib/linux/x86_64_slc6/libgem_hw.so</xc:Module>
   </xc:Context>
 </xc:Partition>
