+++ gemdaq-testing/gemhardware/xml/vfatmanager.xml
@@ -2,8 +2,7 @@
 <xc:Partition xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	      xmlns:soapenc="http://schemas.xmlsoap.org/soap/encoding/"
 	      xmlns:xc="http://xdaq.web.cern.ch/xdaq/xsd/2004/XMLConfiguration-30">
-  <xc:Context url="http://${GEMHOST}:${VFAT_MAN_PORT}">
-
+  <xc:Context url="http://gem904daq01:2222">
     <xc:Application class="gem::hw::vfat::VFAT2Manager" id="3254" instance="1" network="local">
       <properties xmlns="urn:xdaq-application:gem::hw::vfat::VFAT2Manager"
 		  xsi:type="soapenc:Struct">
